requirements.txt
python-dotenv==1.0.1
openai==1.66.0
pydantic==2.10.0
typing-extensions==4.12.2
google-genai==0.3.0
openai-agents==0.0.5
streamlit==1.48.1
google-generativeai
